rename slistScrollAnimation.js
