+++ src/js/helpers/animations/listScrollAnimation.js
